requirements.txt
# Computer Vision and Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.6.0
scikit-image>=0.20.0
scikit-learn>=1.2.0

# Machine Learning and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
gym>=0.26.0

# Geospatial and Satellite Data
folium>=0.14.0
sentinelhub>=3.8.0
rasterio>=1.3.0

# Web and API
requests>=2.28.0
flask>=2.3.0

# Optimization and Hyperparameter Tuning
optuna>=3.0.0

# Data Processing and Analysis
pandas>=2.0.0
scipy>=1.10.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipython>=8.0.0
notebook>=6.5.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=6.0.0

# Utilities
pathlib2>=2.3.7
typing-extensions>=4.5.0
python-dateutil>=2.8.0
tqdm>=4.65.0